requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-Caching==2.1.0
Flask-Compress==1.15
psycopg2-binary==2.9.11
python-dotenv==1.0.1
gunicorn==21.2.0
cryptography>=41.0.0
pandas>=2.2.0
openpyxl>=3.1.2
redis>=5.0.0